% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TRY.R
\name{TRY}
\alias{TRY}
\title{Look for mistakes in dplyr commands}
\usage{
TRY(.data, command)
}
\arguments{
\item{command}{a dplyr command with a piped input}

\item{interactive}{if FALSE, don't ask user whether to continue
when the command has problems}
}
\value{
The output of the dplyr command
}
\description{
When wrapped around a dplyr command, this function
displays the input data frame and scans the command for common
errors such as variable name mis-spellings, failure to name
arguments as appropriate, etc.  If the command is successful,
the command output is returned and it can be assigned, piped to
further operations, etc.
ALL commands must have the data table input PIPED into them.  Don't
use the data table as the first argument (unless it is the right table
in a join operation).
}
\examples{
\dontrun{
  require(mosaicData)
  # height is not a variable in KidsFeet
  # arguments to summarise should be named
  # watch out for na.rm in man
  KidsFeet \%>\% group_by(sex) \%>\% TRY(summarise(mean(height)))
}
}

