% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/editWebFile.R
\name{editWebFile}
\alias{editWebFile}
\title{Display a file from the web in the editor}
\usage{
editWebFile(url)
}
\arguments{
\item{url}{URL locating the file on the web}
}
\value{
The temporary file name, invisibly.
}
\description{
Reads a text file from the web and opens the contents in an editor.
This is useful, for instance, for sharing assignment templates with students.
You MUST "save as" the editor buffer in your own folder
if you want to keep it.
}
\examples{
\dontrun{
   editWebFile("www.mosaic-web.org/go/datasets/angle.R")
}
}

