% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsHelpers.R
\name{scatterGraphHelper}
\alias{DistributionGraphHelper}
\alias{USMap}
\alias{WorldMap}
\alias{barGraphHelper}
\alias{densityGraphHelper}
\alias{distributionGraphHelper}
\alias{scatterGraphHelper}
\title{Interactive interface to ggplot2}
\usage{
scatterGraphHelper(dat)

distributionGraphHelper(data, format = "histogram", default = format,
  system = "ggplot2", show = FALSE, title = "", ...)

barGraphHelper(dat)

densityGraphHelper(...)

USMap(data = NULL, key = NULL, fill = NULL, ...)

WorldMap(data = NULL, key = NULL, fill = NULL, ...)
}
\arguments{
\item{data}{Dataframe containing the variables that might be used in the plot.}

\item{key}{name of variable holding the state or country ID(for mWorldMap and mUSMap only)}

\item{fill}{name of variable to use for the cholopleth map (for mWorldMap and mUSMap only)}
}
\value{
Nothing.  Just for plotting side effects.
}
\description{
Provides a menu selection system (via \code{manipulate}) so that the variables
for different aspects of a plot can be selected interactively.  The \code{ggplot2}
command for generating the plot currently being displayed is copied to the console,
whence it can be copied to a document for later direct, non-interactive use.
}
\examples{
\dontrun{
  scatterGraphHelper(HappinessIndex) # the use menu to map variables to aesthetics
  Counts <-
    Minneapolis2013 \%>\%
    group_by(First, Precinct) \%>\%
    summarise(vote_count=n())
  barGraphHelper(Counts)
  WorldMap(CountryData, key=country, fill=fert)
}
}

