% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representativeCases.R
\name{representativeCases}
\alias{representativeCases}
\title{Select n widely representative random cases}
\usage{
representativeCases(.df, n = 5L, seed = 15)
}
\arguments{
\item{.df}{data frame from which to select rows}

\item{n}{number of rows to return}

\item{seed}{a random seed to use for the random parts of the algorithm.
If you want another result, change the seed.}
}
\description{
Sometimes you want a small sample of a data frame --- as with head() ---
but you want as many different levels of the variables as possible to
be included.  This function takes a sample from a dataframe, arranging things so that
as many distinct levels as possible of each value for each variable 
are kept.  The incoming random seed will be restored when the function exits.
}
\examples{
representativeCases(BabyNames, n=6)
}

