% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRY.R
\name{TRY}
\alias{TRY}
\alias{WAYPOINT}
\alias{WAY_POINT}
\title{Look for mistakes in dplyr commands}
\usage{
TRY(.data, command, interactive = getOption("TRY_interactive", default =
  TRUE))

WAYPOINT(.data, name = NULL, interact = getOption("TRY_interactive", default
  = TRUE))
}
\arguments{
\item{command}{a dplyr command (for a piped input)}

\item{interactive}{if TRUE (the default) ask user whether to continue}

\item{name}{character string naming a waypoint
when the command has problems}
}
\value{
The output of the dplyr command
}
\description{
When wrapped around a dplyr command, TRY()
displays the input data frame and scans the command for common
errors such as variable name mis-spellings, failure to name
arguments as appropriate, etc.  If the command is successful,
the command output is returned and it can be assigned, piped to
further operations, etc.
}
\details{
A WAYPOINT() can be inserted into a chain to suspend the processing
to give you a chance to look at the input and output from the previous
TRY() step. These are displayed in the `TRY INPUT` and `TRY OUTPUT` tabs
in the editor.  Then, you are prompted to continue or quit at this point.

ALL commands must have the data table input PIPED into them.  Don't
use the data table as the first argument (unless it is the right table
in a join operation).
}

