\name{WorldCities}
\alias{WorldCities}
\docType{data}
\title{Cities and their populations}
\description{
A list of cities
}
\usage{data(WorldCities)}
\format{
  A data frame with 23018 observations on the following 10 variables.
  \describe{
    \item{\code{code}}{The ISO (?) city code}
    \item{\code{name}}{Name of the city}
    \item{\code{latitude}}{location in degrees}
    \item{\code{longitude}}{location in degrees}
    \item{\code{country}}{Two letter country code}
    \item{\code{countryRegion}}{A numerical region}
    \item{\code{population}}{Population}
    \item{\code{regionCode}}{ISO (?) Code}
    \item{\code{region}}{Name of the region}
    \item{\code{date}}{Date estimate made}
  }
}
\source{
In Draft: Somewhere on the Internet.  We need a proper source.
}
\references{
In Draft: We need a proper reference
}
\examples{
data(WorldCities)
## maybe str(WorldCities) ; plot(WorldCities) ...
}
\keyword{datasets}
