% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangleApp.R
\name{wrangleApp}
\alias{wrangleApp}
\title{Help with dplyr command construction}
\usage{
wrangleApp(Input)
}
\arguments{
\item{Input}{A data frame or piped input from a chain of \code{dplyr} commands.}
}
\description{
\code{wrangleApp()} opens an interactive (shiny) app that lets you edit
interactively a
potentially complex \code{dplyr} chain of commands. It provides documentation
and argument vetting.
}
\details{
On return, \code{wrangleApp()} opens an editor to display and run the command.  Often,
the user will cut and paste this into a chunk in an Rmd document.
}
\examples{
\dontrun{require(mosaicData)
wrangleApp( KidsFeet \%>\% select( sex, length, width, biggerfoot) )
}
}

